+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/schedule/usecase/QueryEntireSpotScheduleUseCase.kt
@@ -1,7 +1,7 @@
 package team.comit.simtong.domain.schedule.usecase
 
-import team.comit.simtong.domain.schedule.dto.QueryEntireSpotScheduleResponse
-import team.comit.simtong.domain.schedule.dto.SpotScheduleResponse
+import team.comit.simtong.domain.schedule.dto.response.QueryEntireSpotScheduleResponse
+import team.comit.simtong.domain.schedule.dto.response.SpotScheduleResponse
 import team.comit.simtong.domain.schedule.model.Scope
 import team.comit.simtong.domain.schedule.spi.QuerySchedulePort
 import team.comit.simtong.global.annotation.ReadOnlyUseCase
